+++ changelog.md
@@ -9,6 +9,7 @@
     - Make edit context item edit item clicked on instead of first selected item
   - Fix skill window drop down filter not correctly including all item on first load
   - Code editor: Fix remote browse dialog file type filter drop down
+  - Advanced editor: Fix not correctly storing last color when using advanced dialog
 - **Changed:**
   - Update electron 21.4.2 to 21.4.3
   - Update mathjs 11.6.0 to 11.8.1